backend/requirements.txt
accelerate==0.33.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-generator==1.10
async-lru==2.1.0
attrs==25.4.0
babel==2.17.0
beautifulsoup4==4.12.3
bleach==6.3.0
bs4==0.0.2
certifi==2026.1.4
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
courlan==1.3.2
cssselect==1.3.0
cycler==0.12.1
datasets==2.20.0
dateparser==1.2.2
debugpy==1.8.19
decorator==5.2.1
defusedxml==0.7.1
dill==0.3.8
distro==1.9.0
executing==2.2.1
fastapi==0.110.0
fastjsonschema==2.21.2
feedparser==6.0.11
filelock==3.20.1
fonttools==4.61.0
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2024.5.0
geographiclib==2.1
geopandas==1.1.2
geopy==2.4.1
greenlet==3.2.4
groq==1.0.0
grpcio==1.78.0
grpcio-tools==1.78.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
htmldate==1.9.1
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.36.0
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.1
ipykernel==7.1.0
ipython==9.9.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
json5==0.13.0
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.4
jupyterlab==4.5.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
jusText==3.0.2
kiwisolver==1.4.9
lark==1.3.1
lxml==5.2.2
lxml_html_clean==0.4.3
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
mistune==3.2.0
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.16
mypy_extensions==1.1.0
nbclient==0.10.4
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6.1
notebook==7.5.2
notebook_shim==0.2.4
numpy==1.26.4
outcome==1.3.0.post0
overrides==7.7.0
packaging==25.0
Panda3D==1.10.16
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.5
pillow==12.0.0
platformdirs==4.5.1
portalocker==2.10.1
prometheus_client==0.24.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.5
psutil==7.2.1
psycopg2-binary==2.9.11
pure_eval==0.2.3
pyarrow==22.0.0
pyarrow-hotfix==0.7
pycparser==3.0
pydantic==2.6.4
pydantic_core==2.16.3
pygame==2.6.1
Pygments==2.19.2
pyogrio==0.12.1
pyparsing==3.2.5
pypdf==4.2.0
pyproj==3.7.2
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.18.0
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==4.0.0
pytz==2025.2
pywin32==311
pywinpty==3.0.2
PyYAML==6.0.2
pyzmq==27.1.0
qdrant-client==1.9.1
RapidFuzz==3.10.1
readability-lxml==0.8.1
referencing==0.37.0
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.5.1
scipy==1.16.3
selenium==4.40.0
Send2Trash==2.1.0
sentence-transformers==2.6.1
sgmllib3k==1.0.0
shapely==2.1.2
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
SQLAlchemy==2.0.44
stack-data==0.6.3
starlette==0.36.3
sympy==1.14.0
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tld==0.13.1
tokenizers==0.19.1
torch==2.9.1
tornado==6.5.4
tqdm==4.66.5
trafilatura==1.12.2
traitlets==5.14.3
transformers==4.44.2
trio==0.33.0
trio-typing==0.10.0
trio-websocket==0.12.2
types-certifi==2021.10.8.3
types-urllib3==1.26.25.14
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uri-template==1.3.0
urllib3==2.6.3
uvicorn==0.29.0
watchfiles==1.1.1
wcwidth==0.2.14
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==16.0
widgetsnbextension==4.0.15
wsproto==1.3.2
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
